refactor: split dgu join form component
